+++ modules/hw/src/main/java/cz/gattserver/grass3/hw/ui/dialogs/HWDetailsPhotosTab.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 
 import com.vaadin.flow.component.UI;
 import com.vaadin.flow.component.button.Button;
@@ -17,6 +16,8 @@ import cz.gattserver.grass3.hw.HWConfiguration;
 import cz.gattserver.grass3.hw.interfaces.HWItemFileTO;
 import cz.gattserver.grass3.hw.interfaces.HWItemTO;
 import cz.gattserver.grass3.hw.service.HWService;
+import cz.gattserver.grass3.interfaces.UserInfoTO;
+import cz.gattserver.grass3.services.SecurityService;
 import cz.gattserver.grass3.ui.components.OperationsLayout;
 import cz.gattserver.grass3.ui.components.button.DeleteButton;
 import cz.gattserver.grass3.ui.components.button.DetailButton;
@@ -34,8 +35,8 @@ public class HWDetailsPhotosTab extends Div {
 
 	private static final Logger logger = LoggerFactory.getLogger(HWDetailsPhotosTab.class);
 
-	@Autowired
-	private HWService hwService;
+	private transient HWService hwService;
+	private transient SecurityService securityFacade;
 
 	private HWItemTO hwItem;
 	private HWItemDetailsDialog hwItemDetailDialog;
@@ -48,28 +49,42 @@ public class HWDetailsPhotosTab extends Div {
 		init();
 	}
 
+	private HWService getHWService() {
+		if (hwService == null)
+			hwService = SpringContextHelper.getBean(HWService.class);
+		return hwService;
+	}
+
+	private UserInfoTO getUser() {
+		if (securityFacade == null)
+			securityFacade = SpringContextHelper.getBean(SecurityService.class);
+		return securityFacade.getCurrentUser();
+	}
+
 	private void init() {
-		GrassMultiFileBuffer buffer = new GrassMultiFileBuffer();
-
-		Upload upload = new Upload(buffer);
-		upload.addClassName(UIUtils.TOP_MARGIN_CSS_CLASS);
-		upload.setAcceptedFileTypes("image/jpeg", "image/png", "image/gif");
-		upload.addSucceededListener(event -> {
-			try {
-				hwService.saveImagesFile(buffer.getInputStream(event.getFileName()), event.getFileName(), hwItem);
-				populateImages();
-				hwItemDetailDialog.refreshTabLabels();
-			} catch (IOException e) {
-				String msg = "Nezdařilo se uložit obrázek";
-				logger.error(msg, e);
-				new ErrorDialog(msg).open();
-			}
-		});
-
 		containerDiv = new ContainerDiv();
 		containerDiv.setHeight("500px");
 		add(containerDiv);
-		add(upload);
+
+		if (getUser().isAdmin()) {
+			GrassMultiFileBuffer buffer = new GrassMultiFileBuffer();
+			Upload upload = new Upload(buffer);
+			upload.addClassName(UIUtils.TOP_MARGIN_CSS_CLASS);
+			upload.setAcceptedFileTypes("image/jpeg", "image/png", "image/gif");
+			upload.addSucceededListener(event -> {
+				try {
+					getHWService().saveImagesFile(buffer.getInputStream(event.getFileName()), event.getFileName(),
+							hwItem);
+					populateImages();
+					hwItemDetailDialog.refreshTabLabels();
+				} catch (IOException e) {
+					String msg = "Nezdařilo se uložit obrázek";
+					logger.error(msg, e);
+					new ErrorDialog(msg).open();
+				}
+			});
+			add(upload);
+		}
 		populateImages();
 
 		OperationsLayout operationsLayout = new OperationsLayout(e -> hwItemDetailDialog.close());
@@ -82,7 +97,7 @@ public class HWDetailsPhotosTab extends Div {
 		containerDiv.add(gridLayout);
 
 		int counter = 0;
-		for (HWItemFileTO item : hwService.getHWItemImagesFiles(hwItem.getId())) {
+		for (HWItemFileTO item : getHWService().getHWItemImagesFiles(hwItem.getId())) {
 			if (counter == 0)
 				gridLayout.newRow();
 			counter = (counter + 1) % 5;
@@ -95,7 +110,7 @@ public class HWDetailsPhotosTab extends Div {
 
 			Image img = new Image(
 					new StreamResource(item.getName(),
-							() -> hwService.getHWItemImagesFileInputStream(hwItem.getId(), item.getName())),
+							() -> getHWService().getHWItemImagesFileInputStream(hwItem.getId(), item.getName())),
 					item.getName());
 			img.addClassName("thumbnail-200");
 			itemDiv.add(img);
@@ -106,13 +121,16 @@ public class HWDetailsPhotosTab extends Div {
 			DetailButton detailButton = new DetailButton(e -> UI.getCurrent().getPage()
 					.open(HWConfiguration.HW_PATH + "/" + hwItem.getId() + "/img/" + item.getName()));
 			detailButton.getStyle().set("margin-right", "var(--lumo-space-m)");
-			Button delBtn = new DeleteButton(e -> new ConfirmDialog("Opravdu smazat foto HW položky ?", ev -> {
-				hwService.deleteHWItemImagesFile(hwItem.getId(), item.getName());
-				populateImages();
-				hwItemDetailDialog.refreshTabLabels();
-			}).open());
 			buttonLayout.add(detailButton);
-			buttonLayout.add(delBtn);
+
+			if (getUser().isAdmin()) {
+				Button delBtn = new DeleteButton(e -> new ConfirmDialog("Opravdu smazat foto HW položky ?", ev -> {
+					getHWService().deleteHWItemImagesFile(hwItem.getId(), item.getName());
+					populateImages();
+					hwItemDetailDialog.refreshTabLabels();
+				}).open());
+				buttonLayout.add(delBtn);
+			}
 		}
 	}
 }
